# Conflicts:
#	backend-nest/src/question/entity/question.entity.ts
#	backend-nest/src/question/service/question.service.ts
#	backend-nest/src/questionnary/service/questionnary.service.ts

+++ frontend-vue/src/components/questionary/QuestionnaryList.vue
@@ -2,6 +2,12 @@
   import { useQuestionnaryStore } from '@/stores/questionnaryStore';
 
   export default {
+    emits: ['ChangeStatuss'],
+    data() {
+      return {
+        alertQuestionDelete: false,
+      };
+    },
     setup() {
       const useQ = useQuestionnaryStore();
       return {
@@ -15,10 +21,15 @@
     },
     methods: {
       deleteQuest() {
-        this.useQ.deleteQuestion(this.idQuestion);
+        if (!this.alertQuestionDelete) {
+          this.alertQuestionDelete = true;
+        } else {
+          this.alertQuestionDelete = false;
+          this.useQ.deleteQuestion(this.idQuestion);
+        }
       },
       modifyQuest() {
-        this.$emit('ChangeStatuss', this.idQuestion);
+        this.$emit('ChangeStatuss', this.idQuestion, this.typeLabel);
       },
     },
   };
@@ -33,6 +44,21 @@
     <v-btn icon="edit" @click="modifyQuest"></v-btn>
     <v-btn icon="delete" @click="deleteQuest"></v-btn>
   </div>
+
+  <v-dialog v-model="alertQuestionDelete" max-width="600">
+    <v-card>
+      <v-card-title class="headline">Confirmation</v-card-title>
+      <v-card-text>
+        Are you sure to delete the question :
+        <b>" {{ numberLabel }} "</b>
+        ?
+      </v-card-text>
+      <v-card-actions>
+        <v-btn @click="alertQuestionDelete = false">Cancel</v-btn>
+        <v-btn @click="deleteQuest">confirm</v-btn>
+      </v-card-actions>
+    </v-card>
+  </v-dialog>
 </template>
 
 <style scoped>
